Base HRC1155 structure and changes in HRC721 related to BaseToken abstract class.
